13: refactoring code
