--- AdobeStockAsset/Model/OAuth/OAuthException.php
@@ -1,18 +1,0 @@
-<?php
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
-
-declare(strict_types=1);
-
-namespace Magento\AdobeStockAsset\Model\OAuth;
-
-use Magento\Framework\Exception\LocalizedException;
-
-/**
- * Class OAuthException
- */
-class OAuthException extends LocalizedException
-{
-}
