+++ AdobeStockClient/Model/Client.php
@@ -14,17 +14,22 @@ use AdobeStock\Api\Models\SearchParameters;
 use AdobeStock\Api\Models\StockFile;
 use AdobeStock\Api\Request\SearchFiles as SearchFilesRequest;
 use Exception;
+use Magento\AdobeStockAsset\Model\OAuth\OAuthException;
+use Magento\AdobeStockAsset\Model\OAuth\TokenResponse;
 use Magento\AdobeStockClientApi\Api\ClientInterface;
 use Magento\AdobeStockClientApi\Api\SearchParameterProviderInterface;
 use Magento\Framework\Api\AttributeValue;
-use Magento\Framework\Api\Search\DocumentFactory;
-use Magento\Framework\Api\SearchCriteriaInterface;
-use Magento\Framework\Api\Search\SearchResultInterface;
-use Magento\Framework\Api\Search\SearchResultFactory;
 use Magento\Framework\Api\AttributeValueFactory;
+use Magento\Framework\Api\Search\DocumentFactory;
+use Magento\Framework\Api\Search\SearchResultFactory;
+use Magento\Framework\Api\Search\SearchResultInterface;
+use Magento\Framework\Api\SearchCriteriaInterface;
+use Magento\Framework\Exception\AuthorizationException;
 use Magento\Framework\Exception\IntegrationException;
+use Magento\Framework\HTTP\Client\CurlFactory;
 use Magento\Framework\Locale\ResolverInterface as LocaleResolver;
 use Magento\Framework\Phrase;
+use Magento\Framework\Serialize\Serializer\Json;
 use Psr\Log\LoggerInterface;
 
 /**
@@ -68,20 +73,38 @@ class Client implements ClientInterface
     private $connectionFactory;
 
     /**
+     * @var CurlFactory
+     */
+    private $curlFactory;
+
+    /**
+     * @var OAuth\TokenResponseFactory
+     */
+    private $tokenResponseFactory;
+
+    /**
+     * @var Json
+     */
+    private $json;
+
+    /**
      * @var LoggerInterface
      */
     private $logger;
 
     /**
      * Client constructor.
-     * @param Config                           $config
-     * @param DocumentFactory                  $documentFactory
-     * @param SearchResultFactory              $searchResultFactory
-     * @param AttributeValueFactory            $attributeValueFactory
+     * @param Config $config
+     * @param DocumentFactory $documentFactory
+     * @param SearchResultFactory $searchResultFactory
+     * @param AttributeValueFactory $attributeValueFactory
      * @param SearchParameterProviderInterface $searchParametersProvider
-     * @param LocaleResolver                   $localeResolver
-     * @param ConnectionFactory                $connectionFactory
-     * @param LoggerInterface                  $logger
+     * @param LocaleResolver $localeResolver
+     * @param ConnectionFactory $connectionFactory
+     * @param CurlFactory $curlFactory
+     * @param OAuth\TokenResponseFactory $tokenResponseFactory
+     * @param Json $json
+     * @param LoggerInterface $logger
      */
     public function __construct(
         Config $config,
@@ -91,6 +114,9 @@ class Client implements ClientInterface
         SearchParameterProviderInterface $searchParametersProvider,
         LocaleResolver $localeResolver,
         ConnectionFactory $connectionFactory,
+        CurlFactory $curlFactory,
+        OAuth\TokenResponseFactory $tokenResponseFactory,
+        Json $json,
         LoggerInterface $logger
     ) {
         $this->config = $config;
@@ -100,6 +126,9 @@ class Client implements ClientInterface
         $this->searchParametersProvider = $searchParametersProvider;
         $this->localeResolver = $localeResolver;
         $this->connectionFactory = $connectionFactory;
+        $this->curlFactory = $curlFactory;
+        $this->tokenResponseFactory = $tokenResponseFactory;
+        $this->json = $json;
         $this->logger = $logger;
     }
 
@@ -155,6 +184,39 @@ class Client implements ClientInterface
             );
             $this->processException($message, $exception);
         }
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getToken(string $code): OAuth\TokenResponse
+    {
+        $curl = $this->curlFactory->create();
+
+        $curl->addHeader('Content-Type', 'application/x-www-form-urlencoded');
+        $curl->addHeader('cache-control', 'no-cache');
+
+        $curl->post(
+            self::TOKEN_URI,
+            [
+                'client_id' => $this->config->getApiKey(),
+                'client_secret' => $this->config->getPrivateKey(),
+                'code' => $code,
+                'grant_type' => 'authorization_code'
+            ]
+        );
+
+        $tokenResponse = $this->json->unserialize($curl->getBody());
+        $tokenResponse = $this->tokenResponseFactory->create()
+            ->addData(is_array($tokenResponse) ? $tokenResponse : ['error' => 'The response is empty.']);
+
+        if (empty($tokenResponse->getAccessToken()) || empty($tokenResponse->getRefreshToken())) {
+            throw new AuthorizationException(
+                __('Authentication is failing. Error code: %1', $tokenResponse->getError())
+            );
+        }
+
+        return $tokenResponse;
     }
 
     /**
